refactor: split cclib harnesses by native program
