+++ qcengine/programs/cclib_programs/__init__.py
@@ -1,0 +1,7 @@
+"""Extensible cclib-backed native program harnesses."""
+
+from .base import CCLibHarness
+from .cclib_orca import ORCACCLibHarness
+from .cclib_qchem import QChemCCLibHarness
+
+__all__ = ["CCLibHarness", "ORCACCLibHarness", "QChemCCLibHarness"]
